2022/AoC2022.UnitTests/DayTests/Day6Tests.cs
using AoC2022.Days;
using Moq;
using NUnit.Framework;
using System;
using System.Collections.Generic;
using System.Threading.Tasks;
using Utilities;

namespace AoC2022.UnitTests.DayTests;

internal class Day6Tests
{
    [TestCase("mjqjpqmgbljsphdztnvjfqwrcgsmlb", 7)]
    [TestCase("bvwbjplbgvbhsrlpgdmjqwftvncz", 5)]
    [TestCase("nppdvjthqldpwncqszvftbrmjlhg", 6)]
    [TestCase("nznrnfrfntjfmvfwmzdfjlvtqnbhcprsg", 10)]
    [TestCase("zcfzfwzzqfrljwzlrfnpqdbhtmscgvjw", 11)]
    public async Task TestDay6Part1(string input, int expectedResult)
    {
        var inputMock = PrepareInput(input);
        var objUnderTest = new Day6(inputMock.Object);
        Assert.That(await objUnderTest.Part1(), Is.EqualTo(expectedResult));
    }

    [TestCase("mjqjpqmgbljsphdztnvjfqwrcgsmlb", 19)]
    [TestCase("bvwbjplbgvbhsrlpgdmjqwftvncz", 23)]
    [TestCase("nppdvjthqldpwncqszvftbrmjlhg", 23)]
    [TestCase("nznrnfrfntjfmvfwmzdfjlvtqnbhcprsg", 29)]
    [TestCase("zcfzfwzzqfrljwzlrfnpqdbhtmscgvjw", 26)]
    public async Task TestDay6Part2(string input, int expectedResult)
    {
        var inputMock = PrepareInput(input);
        var objUnderTest = new Day6(inputMock.Object);
        Assert.That(await objUnderTest.Part2(), Is.EqualTo(expectedResult));
    }

    private static Mock<InputFetcher> PrepareInput(string input)
    {
        var inputMock = new Mock<InputFetcher> { CallBase = true };
        inputMock.Setup(m => m.FetchInputAsString(It.IsAny<int>())).Returns(Task.FromResult(input));
        return inputMock;
    }
}
